// client/src/styles/colors.ts
enum Colors {
    Text1 = "#ffffff", // Text over Primary1 and Primary2
    Text2 = "#e0e0e0", // Text over Background2
    Text3 = "#c0c0c0", // Text over Background2
    Background1 = "#1e1e1e", // Main background
    Background2 = "#2e2e2e", // Used for cards
    Primary2 = "#8a2be2", // Primary color
    Primary1 = "#9370db", // Secondary primary color
    Accent1 = "#ba55d3", // Accent color
    Accent2 = "#dda0dd", // Secondary accent color
    Danger = "#ff7f7f", // Danger color
}

export default Colors;